backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
openai==1.30.1
langchain==0.2.1
langchain-openai==0.1.8
langchain-community==0.2.1
langgraph==0.1.1
tavily-python==0.3.3
motor==3.4.0
pymongo==4.7.2
chromadb==0.5.0
python-multipart==0.0.9
aiofiles==23.2.1
httpx==0.27.0
pydantic==2.7.1
